--- docs/Contribution.md
@@ -1,25 +1,0 @@
-# Contribution
-
-## Getting the Required Sources
-
-```bash
-# Clone the Repository
-git clone https://github.com/Ciaxur/webhook.js
-```
-
----
-
-## Required Programs
-- A Git program
-    - [Git](https://git-scm.com/) (*Recommended*)
-    - [GitKraken](https://www.gitkraken.com/) 
-- [NodeJS](https://nodejs.org/) (*Latest* `npm` *version installed*)
-
----
-
-## Filing System
-
-A simple neat filing system should be followed in order to stay organized:
-- **`docs`** - Directory used to store documentations
-- **`test`** - Test builds Directory
-- **`resources`** - Project resources such as images
